Watch config file for changes
We can use a `async for changes` statement at the end of our main, so
that we can watch for changes to our config file. This allows us to
automatically watch for changes.

We put a `self.servicelock` around our reading/writing to ensure we do
not get any weirdness if we were to reload a configuration at the same
time a request is getting processed.